Fixed Some Errors
For the easy samples I tried ( CCC, CCCC), the generated Fragments
seemed to be the correct ones, the count was not correct though. I
fragmented Aspirin and generated 24 different Fragments, which is very
close to the figure of 25 in the paper. So I guess Im on the right
track, but need to put in more work to make sure it is actually working
correct.